+++ world/world.lua
@@ -4,6 +4,7 @@ local dungeonGen = require("world.dungeonGen")
 local populate = require("world.populate")
 local constants = require("constants")
 local tarotMod = require("systems.tarot")
+local mutatorsMod = require("systems.roomMutators")
 
 -- Tracked state
 world.etapa = 1
@@ -64,6 +65,7 @@ function world.init()
     world.dungeon = nil
     world.generarMazmorra()
     tarotMod.reset()
+    mutatorsMod.resetStage()
     coreWorld.set("mundoCompletado", false)
 end
 
@@ -209,6 +211,7 @@ function world.avanzarEtapa()
     world.etapa = world.etapa + 1
     world.generarMazmorra()
     tarotMod.reset()
+    mutatorsMod.resetStage()
     world.puntajeSala = 0
     world.objetivoSala = world.calcularObjetivo()
 end